requirements.txt
asgiref>=3.8.1
build>=1.2.2.post1
click>=8.1.8
Django>=5.1.7
django-cors-headers>=4.7.0
djangorestframework>=3.15.2
et_xmlfile>=2.0.0
numpy>=2.2.4
openpyxl>=3.1.5
packaging>=24.2
pandas>=2.2.3
psycopg2-binary>=2.9.10
pyproject_hooks>=1.2.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2025.1
setuptools>=75.8.2
six>=1.17.0
sqlparse>=0.5.3
tzdata>=2025.1
wheel>=0.45.1